+++ package.json
@@ -58,6 +58,7 @@
     "ramda": "^0.27.1",
     "semver": "^7.3.4",
     "strip-ansi": "^6.0.0",
+    "tencent-serverless-http": "^1.3.1",
     "traverse": "^0.6.6",
     "uuid": "^8.3.2"
   },
